chore: move logger to test runner cli
