+++ .changeset/shiny-elephants-shout.md
@@ -1,0 +1,9 @@
+---
+'@web/dev-server-core': patch
+'@web/test-runner-cli': patch
+'@web/test-runner-core': patch
+'@web/test-runner-server': patch
+'@web/test-runner': patch
+---
+
+move logger to test runner cli
